requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.1
argparse==1.2.1
boto==2.9.6
distribute==0.6.31
dj-database-url==0.2.1
django-adaptors==0.2.5
django-facebook==5.2.6
django-grappelli==2.4.5
django-storages==1.1.8
django-tastypie==0.9.15
djorm-ext-pgfulltext==0.8
fb==0.4.0
gunicorn==0.17.4
lxml==3.2.3
mimeparse==0.1.3
oauthlib==0.6.0
psycopg2==2.5
pyes==0.20.0
python-dateutil==2.1
requests==2.0.0
requests-oauthlib==0.3.2
six==1.3.0
twython==3.1.0
urllib3==1.6
wsgiref==0.1.2
